code review tidyup (moving things around)
